+++ FinancialFreeApps/ChartTableViewCell.swift
@@ -10,12 +10,28 @@ import UIKit
 
 class ChartTableViewCell: UITableViewCell {
 
+    var appLink: String?
+
     @IBOutlet weak var appIcon: UIImageView!
+    @IBOutlet weak var rank: UILabel!
     @IBOutlet weak var appName: UILabel!
+    @IBOutlet weak var installButton: UIButton!
+    @IBAction func didTapInstall(_ sender: Any) {
+        if let url = URL(string: self.appLink ?? "") {
+            UIApplication.shared.open(url, options: [:], completionHandler: nil)
+        }
+    }
     
     override func awakeFromNib() {
         super.awakeFromNib()
         // Initialization code
+        appIcon.layer.cornerRadius = 12
+        appIcon.layer.borderWidth = 1
+        appIcon.layer.borderColor = UIColor(red: 0.8, green: 0.8, blue: 0.8, alpha: 0.5).cgColor
+        appIcon.clipsToBounds = true
+        
+        installButton.layer.cornerRadius = 15
+        
     }
 
     override func setSelected(_ selected: Bool, animated: Bool) {
