--- FinancialFreeApps/MasterViewController.swift
@@ -1,119 +1,0 @@
-//
-//  MasterViewController.swift
-//  FinancialFreeApps
-//
-//  Created by Samuel Kim on 2018. 1. 22..
-//  Copyright © 2018년 Samuel Kim. All rights reserved.
-//
-
-import UIKit
-import SwiftyJSON
-
-class MasterViewController: UITableViewController {
-
-    var detailViewController: DetailViewController? = nil
-    var apps = [AppModel]()
-
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        DispatchQueue.global().async {
-            let urlString = "https://itunes.apple.com/kr/rss/topfreeapplications/limit=50/genre=6015/json"
-            if let url = URL(string: urlString) {
-                URLSession.shared.dataTask(with: url) { [weak self] data, response, error in
-                    guard let data = data else {
-                        return
-                    }
-                    if let json = try? JSON(data: data) {
-                        for item in json["feed"]["entry"].arrayValue {
-                            let name = item["im:name"]["label"].string!
-                            let iconUrl = item["im:image"].arrayValue[2]["label"].string!
-                            let link = item["link"]["attributes"]["href"].string!
-                            self?.apps.append(AppModel(name: name, iconUrl: iconUrl, downloadUrl: link))
-                        }
-                    }
-                    DispatchQueue.main.async {
-                        self?.tableView.reloadData()
-                    }
-                }.resume()
-            }
-        }
-        // Do any additional setup after loading the view, typically from a nib.
-        navigationItem.leftBarButtonItem = editButtonItem
-
-//        let addButton = UIBarButtonItem(barButtonSystemItem: .add, target: self, action: #selector(insertNewObject(_:)))
-//        navigationItem.rightBarButtonItem = addButton
-        if let split = splitViewController {
-            let controllers = split.viewControllers
-            detailViewController = (controllers[controllers.count-1] as! UINavigationController).topViewController as? DetailViewController
-        }
-    }
-
-    override func viewWillAppear(_ animated: Bool) {
-        clearsSelectionOnViewWillAppear = splitViewController!.isCollapsed
-        super.viewWillAppear(animated)
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-
-//    @objc
-//    func insertNewObject(_ sender: Any) {
-//        apps.insert(NSDate(), at: 0)
-//        let indexPath = IndexPath(row: 0, section: 0)
-//        tableView.insertRows(at: [indexPath], with: .automatic)
-//    }
-
-    // MARK: - Segues
-
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        if segue.identifier == "showDetail" {
-            if let indexPath = tableView.indexPathForSelectedRow {
-                let object = apps[indexPath.row] as! AppModel
-                let controller = (segue.destination as! UINavigationController).topViewController as! DetailViewController
-                controller.detailItem = object
-                controller.navigationItem.leftBarButtonItem = splitViewController?.displayModeButtonItem
-                controller.navigationItem.leftItemsSupplementBackButton = true
-            }
-        }
-    }
-
-    // MARK: - Table View
-
-    override func numberOfSections(in tableView: UITableView) -> Int {
-        return 1
-    }
-
-    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return apps.count
-    }
-
-    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: "Cell", for: indexPath)
-
-        let object = apps[indexPath.row] as! AppModel
-        cell.textLabel!.text = object.name
-        return cell
-    }
-
-    override func tableView(_ tableView: UITableView, canEditRowAt indexPath: IndexPath) -> Bool {
-        // Return false if you do not want the specified item to be editable.
-        return true
-    }
-
-    override func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCellEditingStyle, forRowAt indexPath: IndexPath) {
-        if editingStyle == .delete {
-            apps.remove(at: indexPath.row)
-            tableView.deleteRows(at: [indexPath], with: .fade)
-        } else if editingStyle == .insert {
-            // Create a new instance of the appropriate class, insert it into the array, and add a new row to the table view.
-        }
-    }
-
-    
-
-}
-
